refactor get_summary_statistics
The previous approach for evaluating summary statistics (median,
uncertainty intervals, etc) was a weighted sampling, with replacement,
from the posterior PDF.

While this approach is good enough for most use cases, as-implemented,
it introduced quantization in the resulting summary statistics at the
level of the age grid.  In other words, if the user's grid resolution
was 1 Myr, the maximum numerical precision they would achieve would be 1
Myr.  This is of course was more precise than any realistic gyro age
uncertainty.

This commit circumvents the grid quantization by interpolating over the
posterior probability density function, computing the CDF, and
iteratively numerically bracketing the CDF to find the median and
uncertainty intervals.