models: add time-variable gaussian scatter
Differential rotation probably imposes a ~5-10% floor on the scatter
we can ever expect to see in the calibration sequences.

In BPH2023, we already had some evidence that this was happening in the
Ruprecht-147 sequence.  Now, with the M67 data from Dungee et al., and
from Gruner et al., it really makes sense to incorporate this increase
in the assumed gaussian scatter in the calibration populations.

That is what this commit does.

(There is also a minor bugfix that yields better nan-handlings for Teff
values outside the 3800-6200K range).